started toc.
HOWEVER, for the basic activity, I think no sections should be shown, it should just list the activity.

When we do \part (as a chapter heading... should we use \chapter for this??)
And follow with \sectionstyle, things look good -- Be sure to check out PDF table of contents (given by PDF viewer)

On the other hand, there is something wrong with chapterstyle... to see this, just try to do

\setcounter{tocdepth}{1}

and note that the style there changes...

This should not do that!